+++ test/Option.ts
@@ -1,4 +1,4 @@
-import { unsafeUnwrap } from "../src/Option"
+import { unsafeUnwrap, noneAs } from "../src/Option"
 import * as O from "fp-ts/Option"
 
 describe("Option", () => {
@@ -11,6 +11,14 @@ describe("Option", () => {
 
     it("throws None", () => {
       expect(() => f(O.none)).toThrow()
+    })
+  })
+
+  describe("noneAs", () => {
+    const f = noneAs
+
+    it("is identical to standard None constructor at runtime", () => {
+      expect(f<unknown>()).toEqual(O.none)
     })
   })
 })